requirements.txt
requests
python-chess
gTTS
moviepy
cairosvg
pillow